+++ engine/src/objects/physics/colliders.rs
@@ -1,8 +1,6 @@
 //! Wrapping of Rapiers colliders to be used with Let Engine and Glam.
 
-use super::VHACDParameters;
 use crate::prelude::*;
-use glam::Vec2;
 use rapier2d::prelude::*;
 
 #[derive(Clone)]